requirements.txt
asgiref==3.8.1
Django
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytils==0.4.1
sqlparse==0.5.0
tzdata==2024.1
flake8==7.0.0
pyflakes==3.2.0
mccabe==0.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
coverage==7.5.3
drf-yasg==1.21.7
stripe==9.12.0
celery==5.4.0
redis==5.0.6
eventlet==0.36.1
django-celery-beat==2.6.0